Cleaned up and rewrote runtest.in to support multiple arguments.
This patch rewrote runtest.in to support multiple arguments in each line
of testline. It checks following format for each line (new argument can
be added in future):
<program> <exec=1|empty by default>

For the second exec flag (exec=1|0), it is equal to 1 by default if the
item is empty. For a test which is enabled by special configure option,
it can be set as exec=@MACRO_FLAG@.