Thread-safe version of error handler fix.
To ensure correctness of error handler in multi-threaded MPI program, we
need to make the external error object flag be in thread local storage.
Because pthread_set{get}specific may not work correctly with OpenMP
(implementation dependent), we do not use it in Casper for portability.
Instead, we use compiler TLS specifier (e.g., gcc provides `__thread`)
if configure detected it is supported, otherwise we build a hash table
to store per-thread TLS data.